% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espnscrapeR-package.R
\docType{package}
\name{espnscrapeR-package}
\alias{espnscrapeR}
\alias{espnscrapeR-package}
\title{espnscrapeR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
