% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_538_elo.R
\name{get_538_elo}
\alias{get_538_elo}
\title{Get ELO ratings from FiveThirtyEight}
\usage{
get_538_elo(season = 2020, stat = "weekly_elo")
}
\arguments{
\item{season}{character or numeric, must be 2015 or greater}

\item{stat}{character, must be one of "games", "weekly_elo", "weekly_rating", "distance", "qb_adj_playoff", "qb_adj"}
}
\value{
Returns a tibble
}
\description{
Get ELO ratings from FiveThirtyEight
}
\examples{
# Get elo for specific season and stat
get_538_elo(season = 2020, stat = "weekly_elo")
}
