% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_538_elo.R
\name{get_538_elo_historical}
\alias{get_538_elo_historical}
\title{Get historical ELO ratings from FiveThirtyEight}
\usage{
get_538_elo_historical(team = "pit")
}
\arguments{
\item{team}{character, team abbreviated name, must be one of "ari", "atl", "bal", "buf", "car", "chi", "cin", "dal", "den", "det", "gb", "hou", "ind", "jax", "kc", "oak", "lac", "lar", "mia", "min", "ne", "no", "nyg", "nyj", "phi", "pit", "sf", "sea", "tb", "ten", "wsh"}
}
\value{
Returns a tibble
}
\description{
Get historical ELO ratings from FiveThirtyEight
}
\examples{

# Get historic elo for specific team
get_538_elo_historical(team = "pit")
}
