% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete.R
\name{get_athlete}
\alias{get_athlete}
\title{Get ESPN athlete info for NFL players}
\usage{
get_athlete(athlete_id)
}
\arguments{
\item{athlete_id}{The player's unique athlete id}
}
\value{
Returns a tibble
}
\description{
Get ESPN athlete info for NFL players
}
\examples{
# Get ALL Playoff QBR from 2016 season
get_athlete("2580")

}
