% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_college_qbr.R
\name{get_college_qbr}
\alias{get_college_qbr}
\title{Get ESPN QBR for College football}
\usage{
get_college_qbr(season = 2020, type = "season")
}
\arguments{
\item{season}{Numeric or character - greater than 2004}

\item{type}{character - "season" or "weekly"}
}
\value{
tibble
}
\description{
Get ESPN QBR for College football
}
\examples{

# Get college QBR from 2014 for players at season level
get_college_qbr(season = 2014, type = "season")

# Get weekly college QBR from 2019
# Note the mix of playoffs/bowls and regular season
get_college_qbr(2019, type = "weekly")
}
