% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_college_espn_qbr.R
\name{get_espn_college_qbr}
\alias{get_espn_college_qbr}
\title{Get ESPN QBR for College football}
\usage{
get_espn_college_qbr(season = 2019, week = NA)
}
\arguments{
\item{season}{Numeric or character - greater than 2004}

\item{week}{Numeric or character - typically 1 to 15 or Bowls}
}
\value{
tibble
}
\description{
Get ESPN QBR for College football
}
\examples{

get_college_qbr(season = 2011, week = NA)

get_espn_college_qbr(2019, 1)

}
