% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_api_standings.R
\name{get_espn_nfl_standings}
\alias{get_espn_nfl_standings}
\title{Get NFL standings for a specific season from ESPN's API}
\usage{
get_espn_nfl_standings(season = 2019)
}
\arguments{
\item{season}{Either numeric or character}
}
\value{
Returns a tibble
}
\description{
Get NFL standings for a specific season from ESPN's API
}
\examples{
get_espn_nfl_standings(season = "2018")

get_espn_nfl_standings(2010)

}
