% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_espn_standings.R
\name{get_espn_standings}
\alias{get_espn_standings}
\title{Collect NFL standings for a specific season from ESPN's API}
\usage{
get_espn_standings(season = 2019)
}
\arguments{
\item{season}{Either numeric or character}
}
\value{
Returns a tibble
}
\description{
Collect NFL standings for a specific season from ESPN's API
}
