% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_odds.R
\name{get_nfl_odds}
\alias{get_nfl_odds}
\title{Get NFL odds for a specific game from ESPN's API}
\usage{
get_nfl_odds(game_id)
}
\arguments{
\item{game_id}{character}
}
\value{
Returns a tibble
}
\description{
Get NFL odds for a specific game from ESPN's API
}
\examples{
# Get odds from a specific game
get_nfl_odds(game_id = "400791550")
}
