% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_pbp.R
\name{get_nfl_pbp}
\alias{get_nfl_pbp}
\title{Get NFL play-by-play for a specific game from ESPN's API}
\usage{
get_nfl_pbp(game_id)
}
\arguments{
\item{game_id}{character}
}
\value{
Returns a tibble
}
\description{
Get NFL play-by-play for a specific game from ESPN's API
}
\examples{
# Get NFL play-by-play for a specific game
get_nfl_pbp(game_id = "300912027")
}
