% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nfl_espn_qbr.R
\name{get_nfl_qbr}
\alias{get_nfl_qbr}
\title{Get ESPN QBR for NFL football}
\usage{
get_nfl_qbr(season = 2019, week = NA, season_type = "Regular")
}
\arguments{
\item{season}{Either numeric or character}

\item{week}{Either NA to return season or week 1 to 4 for playoffs or 1 to 17 for regular}

\item{season_type}{Character - either "Regular" or "Playoffs"}
}
\value{
Returns a tibble
}
\description{
Get ESPN QBR for NFL football
}
\examples{
get_nfl_qbr("2016", season_type = "Playoffs", week = NA)

all_superbowls <- tidyr::crossing(season = 2006:2019, season_type = "Playoffs", week = 4) \%>\%
dplyr::filter(season != 2017) \%>\%
purrr::pmap_dfr(get_nfl_qbr)

get_nfl_qbr("2019", season_type = "Regular", week = 4)

}
