% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_qbr_helper.R
\name{get_nfl_qbr_helper}
\alias{get_nfl_qbr_helper}
\title{Get ESPN QBR for NFL football}
\usage{
get_nfl_qbr_helper(raw_json, week, season, season_type)
}
\arguments{
\item{week}{Either NA to return season or week 1 to 4 for playoffs or 1 to 17 for regular}

\item{season}{Either numeric or character}

\item{season_type}{Character - either "Regular" or "Playoffs"}
}
\value{
Returns a tibble
}
\description{
Get ESPN QBR for NFL football
}
