% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_schedule.R
\name{get_nfl_schedule}
\alias{get_nfl_schedule}
\title{Get NFL schedule for a specific year from ESPN's API}
\usage{
get_nfl_schedule(season)
}
\arguments{
\item{season}{Either numeric or character}
}
\value{
Returns a tibble
}
\description{
Get NFL schedule for a specific year from ESPN's API
}
\examples{
# Get all games from 2018 season, note that this will have some overlap
# between seasons, for example 2018 returns 2018-01-01 to 2018-12-31
get_nfl_schedule(season = "2018")
}
