% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_standings.R
\name{get_nfl_standings}
\alias{get_nfl_standings}
\title{Get NFL standings for a specific season from ESPN's API}
\usage{
get_nfl_standings(season = 2019, quiet = FALSE)
}
\arguments{
\item{season}{Either numeric or character}
}
\value{
Returns a tibble
}
\description{
Get NFL standings for a specific season from ESPN's API
}
\examples{
# Get standings from 2018 season
get_nfl_standings(season = "2018")

# Get standings from 2010 season
get_nfl_standings(2010)


}
