% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_teams.R
\name{get_nfl_teams}
\alias{get_nfl_teams}
\title{Return all NFL teams with their name, abbreviation,}
\usage{
get_nfl_teams()
}
\value{

}
\description{
Return all NFL teams with their name, abbreviation,
}
\examples{

get_nfl_teams()
}
