% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sharpe_data.R
\name{get_sharpe_data}
\alias{get_sharpe_data}
\title{Get some of Lee Sharpe's datasets from GitHub}
\usage{
get_sharpe_data(dataset = "games")
}
\arguments{
\item{dataset}{Character only, one of "airports", "closing_lines", "draft_picks", "draft_values", "games", "logos", "pff_pfr_map_v1", "positions", "rosters", "sc_lines", "standings", "teamcolors", "teams", "trades", "win_totals"}
}
\value{
tibble
}
\description{
Get some of Lee Sharpe's datasets from GitHub
}
\examples{

# Get NFL games and their outcomes
get_sharpe_data(dataset = "games")

# Get team abbreviations/names/etc by year
get_sharpe_data(dataset = "teams")
}
