% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_hulk_color.R
\name{gt_hulk_color}
\alias{gt_hulk_color}
\title{Apply 'hulk' palette to specific columns in a gt table.}
\usage{
gt_hulk_color(gt_object, ..., domain = NULL)
}
\arguments{
\item{gt_object}{An existing gt table object}

\item{...}{columns to apply color to}
}
\value{
Returns a gt table
}
\description{
Apply 'hulk' palette to specific columns in a gt table.
}
