% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_538.R
\name{gt_theme_538}
\alias{gt_theme_538}
\title{Apply FiveThirtyEight theme to a gt table}
\usage{
gt_theme_538(data, ...)
}
\arguments{
\item{data}{An existing gt table object}

\item{...}{Optional additional arguments to gt::table_options()}
}
\value{
Returns a tibble
}
\description{
Apply FiveThirtyEight theme to a gt table
}
