% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_espn.R
\name{gt_theme_espn}
\alias{gt_theme_espn}
\title{Apply ESPN theme to a gt table}
\usage{
gt_theme_espn(data, ...)
}
\arguments{
\item{data}{An existing gt table object}

\item{...}{Optional additional arguments to gt::table_options()}
}
\value{
Returns a tibble
}
\description{
Apply ESPN theme to a gt table
}
