% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_espn_stats.R
\name{scrape_espn_stats}
\alias{scrape_espn_stats}
\title{Scrape NFL stats from ESPN}
\usage{
scrape_espn_stats(season = 2019, stats = "receiving", season_type = "Regular")
}
\arguments{
\item{season}{character or numeric - greater than 1990}

\item{stats}{character - either receiving, passing, or rushing}

\item{season_type}{character - either Regular or Playoffs}
}
\value{
tibble
}
\description{
Scrape NFL stats from ESPN
}
\examples{
scrape_espn_stats(season = 2000, stats = "passing")

}
