% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_espn_win_rate.R
\name{scrape_espn_win_rate}
\alias{scrape_espn_win_rate}
\title{Scrape ESPN Pass/Run Block/Rush Win Rates ratings for a specific season from ESPN's site}
\usage{
scrape_espn_win_rate(season = 2020)
}
\value{
Returns a tibble
}
\description{
Scrape ESPN Pass/Run Block/Rush Win Rates ratings for a specific season from ESPN's site
}
\examples{
# Get off and def pass/run win rates
scrape_espn_win_rate()
}
