% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fpi.R
\name{scrape_fpi}
\alias{scrape_fpi}
\title{Scrape ESPN FPI ratings for a specific season from ESPN's site}
\usage{
scrape_fpi(season = 2020, stat = "FPI")
}
\arguments{
\item{season}{Either numeric or character}

\item{stat}{One of 'FPI', 'EFF' or 'PROJ'}
}
\value{
Returns a tibble
}
\description{
Scrape ESPN FPI ratings for a specific season from ESPN's site
}
\examples{
# Get team offensive and defensive efficiency from 2020 season
scrape_fpi(2020, stat = "EFF")
}
