% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nfl_qbr.R
\name{scrape_nfl_qbr}
\alias{scrape_nfl_qbr}
\title{Scrape ESPN QBR for NFL football}
\usage{
scrape_nfl_qbr(season = 2019, week = NA, season_type = "Regular")
}
\arguments{
\item{season}{Either numeric or character. Must be between 2006 and current season.}

\item{week}{Either NA to return season, week 1 to 4 for playoffs, or 1 to 17 for regular season.}

\item{season_type}{Character - either "Regular" or "Playoffs"}
}
\value{
Returns a tibble
}
\description{
Scrapes QBR tables for NFL QBs by specific week or at season level.
}
\details{
This function supercedes `get_nfl_qbr` due to the ESPN API altering past data.
}
\examples{
# Scrape ALL Playoff QBR from 2016 season
scrape_nfl_qbr("2016", season_type = "Playoffs", week = NA)

# Scrape Regular season QBR for week 4 of 2019
scrape_nfl_qbr("2019", season_type = "Regular", week = 4)
}
