% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nfl_standings.R
\name{scrape_nfl_standings}
\alias{scrape_nfl_standings}
\title{Scrape NFL standings for a specific season from ESPN's site}
\usage{
scrape_nfl_standings(season, add_superbowls = FALSE)
}
\arguments{
\item{season}{Either numeric or character}

\item{add_superbowls}{Join superbowl winners (appropriate for historical data or after season over)}
}
\value{
Returns a tibble
}
\description{
Scrape NFL standings for a specific season from ESPN's site
}
\examples{
# Get standings from 2018 season
scrape_nfl_standings(season = "2018")

# Get standings from 2010 season
scrape_nfl_standings(2000, add_superbowls = TRUE)
}
