% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nfl_weekly_standings.R
\name{scrape_nfl_weekly_standings}
\alias{scrape_nfl_weekly_standings}
\title{Scrape NFL weekly outcomes by week}
\usage{
scrape_nfl_weekly_standings(season = 2020, tidy = FALSE)
}
\arguments{
\item{season}{character or numeric - greater than 1990}

\item{tidy}{logical - either TRUE to stack data by game/week or FALSE to return table as is}
}
\value{
tibble
}
\description{
Scrape NFL weekly outcomes by week
}
\examples{

# Here we run w/ tidy = FALSE to get the exact table from PFR
scrape_nfl_weekly_standings(season = 2020, tidy = FALSE)

# Here we scrape the outcome and stack the games on top of eachother
scrape_nfl_weekly_standings(season = 2020, tidy = TRUE)
}
