% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_superbowls.R
\name{scrape_superbowls}
\alias{scrape_superbowls}
\title{Scrape NFL superbowl winners for all years}
\usage{
scrape_superbowls()
}
\value{
Returns a tibble
}
\description{
Scrape NFL superbowl winners for all years
}
\examples{
# Get superbowls
scrape_superbowls()
}
