% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_team_stats_nfl.R
\name{scrape_team_stats_nfl}
\alias{scrape_team_stats_nfl}
\title{Scrape NFL stats from nfl.com}
\usage{
scrape_team_stats_nfl(
  season = 2019,
  stats = "GAME_STATS",
  role = "offense",
  season_type = "Regular"
)
}
\arguments{
\item{season}{character or numeric - greater than 1970}

\item{stats}{character - either "GAME_STATS", "SCORING", "TEAM_PASSING", "TEAM_RUSHING", "TEAM_RECEIVING", or "OFFENSIVE_LINE"}

\item{role}{character - "offense" or "defense"}

\item{season_type}{character - either 'Regular' or 'Playoffs'}
}
\value{
tibble
}
\description{
Scrape NFL stats from nfl.com
}
\examples{
scrape_team_stats_nfl(season = 2018, stats = "GAME_STATS", role = "offense")
}
