% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espnscrapeR-package.R
\docType{package}
\name{espnscrapeR-package}
\alias{espnscrapeR}
\alias{espnscrapeR-package}
\title{espnscrapeR: Scrapes Or Collects NFL Data From ESPN}
\description{
Main use case is to collect ESPN QBR for NFL and college football. Alternative functions include getting NFL standings and scraping NFL season-level stats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jthomasmock/espnscrapeR}
  \item Report bugs at \url{https://github.com/jthomasmock/espnscrapeR/issues}
}

}
\keyword{internal}
