% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth_chart.R
\name{get_depth_chart}
\alias{get_depth_chart}
\title{Get ESPN NFL depth chart by year and team}
\usage{
get_depth_chart(season = 2020, team = 23)
}
\arguments{
\item{season}{Either numeric or character}

\item{team}{team_id, can be retrieved via get_nfl_teams}
}
\value{
Returns a tibble
}
\description{
Get ESPN NFL depth chart by year and team
}
\examples{
# Get ALL Playoff QBR from 2016 season
get_depth_chart("2016", team = 23)

}
