% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfl_schedule.R
\name{get_nfl_schedule}
\alias{get_nfl_schedule}
\title{Get NFL schedule for a specific year from ESPN's API}
\usage{
get_nfl_schedule(season)
}
\arguments{
\item{season}{Either numeric or character}
}
\value{
Returns a tibble
}
\description{
Get NFL schedule for a specific year from ESPN's API
}
\examples{
# Get standings from 2018 season
get_nfl_schedule(season = "2018")
}
