% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_team_stats_nfl.R
\name{scrape_team_stats_nfl}
\alias{scrape_team_stats_nfl}
\title{Scrape NFL stats from nfl.com at the team level.}
\usage{
scrape_team_stats_nfl(season = 2019, stats = "passing", role = "offense")
}
\arguments{
\item{season}{character or numeric - greater than 1971, note that some data is missing for earlier seasons.}

\item{stats}{character - either "passing", "rushing", "receiving", "scoring", "downs"}

\item{role}{character - "offense" or "defense"}
}
\value{
tibble
}
\description{
Please note that the column names are identical between offense/defense and you should reference the 'role' column for offensive or defensive stats.
}
\examples{

# Get the NFL.com game-level stats for offense in 2018
scrape_team_stats_nfl(season = 2018, stats = "passing", role = "offense")

# Get the NFL.com team-level passing stats for defense in 2014
scrape_team_stats_nfl(season = "2014", stats = "rushing", role = "defense")
}
