% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_weekly_leaders.R
\name{scrape_weekly_leaders}
\alias{scrape_weekly_leaders}
\title{Scrape NFL stats from ESPN}
\usage{
scrape_weekly_leaders(
  season = 2019,
  week = 1,
  season_type = "Regular",
  stats = "passing"
)
}
\arguments{
\item{season}{character or numeric - greater than 1990}

\item{season_type}{character - either Regular or Playoffs}

\item{stats}{character - either receiving, passing, or rushing}
}
\value{
tibble
}
\description{
Scrape NFL stats from ESPN
}
\examples{
scrape_espn_stats(season = 2000, stats = "passing")
}
